dashboard: avoid Django and disk write if possible
If the parameters given to the Django template (and later written to
disk) are unchanged, it's absolutely unnecessary to import Django, parse
the template, render the unchanged response, and write the same content
to the disk that was stored before.

With this change, the whole Django import - render - disk write chain
happens only if the SHA-1 checksum of the serialized data doesn't
match the one stored at the last rendering, the output file doesn't
exist, or this is the first rendering.